// src/renderer/types/action-types.d.ts
export const enum ActionType {
    TransitionIn = -2,
    TransitionOut = -1,
    TweenAnimation = 0,
    EffectAnimation = 1,
    ModifyVariable = 2,
    PlayAudio = 3,
    EmitEvent = 4,
    TextAnimation = 5
}

export const enum TriggerType {
    PreviousStart,
    PreviousEnd,
    Click,
    Event
}

export interface ActionBase {
    name: string;
    type: ActionType;
    trigger: TriggerType;
    event?: string;
    delay: number;
}

export const enum VariableOperator {
    Assign = 1,
    Plus = 2,
    Minus = 3,
    Multiply = 4,
    Divide = 5,
    Modulo = 6
}

export const enum OperandType {
    Variable = 1,
    Const = 2,
    Random = 3
}
// ...
export interface SimpleValueTeenAnimation extends TweenAnimationBase {
    propertyName: "rotation" | "opacity";
    value: number
}

export interface ScaleTweenValueType {
    scaleX: number, scaleY: number
}

export interface ScaleTweenAnimation extends TweenAnimationBase {
    propertyName: "scale";
    value: ScaleTweenValueType;
}
// ...
export interface PositionTweenAnimation extends TweenAnimationBase {
    propertyName: "position";
    value: PositionTweenValueType;
}

export type TweenAnimation = PositionTweenAnimation | SimpleValueTeenAnimation | ScaleTweenAnimation;
// ...
export interface TextAnimation extends ActionBase {
    type: ActionType.TextAnimation;
    target: string;
    speed: number;
}

export type Action = ModifyVariableAction | TweenAnimation | PlayAudioAction | EmitEventAction | TextAnimation;
// ...
export type TransitionInAction = NoneTransitionInAction | FadeTransitionInAction | MoveTransitionInAction;
// ...
export type TransitionOutAction = NoneTransitionOutAction | DurationTransitionOutAction | MoveTransitionOutAction;

export interface Actions {
    transitionIn: TransitionInAction;
    transitionOut: TransitionOutAction;
    actions: Action[];
}

export interface DialogueAction {
    [key: string]: Actions;
}